+++ runtime/execdriver/native/driver.go
@@ -3,9 +3,9 @@ package native
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/dotcloud/docker/pkg/apparmor"
 	"github.com/dotcloud/docker/pkg/cgroups"
 	"github.com/dotcloud/docker/pkg/libcontainer"
-	"github.com/dotcloud/docker/pkg/libcontainer/apparmor"
 	"github.com/dotcloud/docker/pkg/libcontainer/nsinit"
 	"github.com/dotcloud/docker/pkg/system"
 	"github.com/dotcloud/docker/runtime/execdriver"
